+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Changed
+- `cassie migrations:import` now dumps the schema after succeeding
+
+### Fixed
+- bug where `cassie:migrations:import` didn't initialize schema
+- bug where `Cassie::Schema.record_version` silently failed
+
 ## 1.0.3
 
 ### Fixed